Fix us leaving the scanner in waiting state until timeout
...after downloading the scanned page. My Brother MFC-L2710DW stayed in
"PC connection mode" after we downloaded the page until it timed out.
So it seemed to not consider the scan job done. Looking at what macOS
does via eSCL (and the recently released spec), it turned out to expect
another `NextDocument` request that it would answer with HTTP 404. This
completes the job properly and leaves the scanner ready for the next
job.